Reorganized and merged authentication feature folder
