+++ lib/core/values/app_imports.dart
@@ -19,7 +19,9 @@ export 'package:keno_plus/features/home/presentation/widgets/home_widgets.dart';
 export 'package:keno_plus/features/loading_screen/presentation/screens/loading_screen.dart';
 export 'package:keno_plus/features/loading_screen/presentation/widgets/loading_screen_widgets.dart';
 export 'package:keno_plus/features/profile/presentation/screen/profile.dart';
-export 'package:keno_plus/features/profile/authentication/presentation/bloc/auth_bloc.dart';
-export 'package:keno_plus/features/profile/authentication/presentation/widgets/auth_widgets.dart';
-export 'package:keno_plus/features/profile/authentication/presentation/widgets/login_widget.dart';
-export 'package:keno_plus/features/profile/authentication/presentation/widgets/signup_widget.dart';
+export 'package:keno_plus/features/authentication/presentation/bloc/authentication_bloc.dart';
+export 'package:keno_plus/features/authentication/presentation/screens/log_in.dart';
+export 'package:keno_plus/features/authentication/presentation/screens/sign_up.dart';
+export 'package:keno_plus/features/authentication/presentation/widgets/authentication_widgets.dart';
+export 'package:keno_plus/features/authentication/presentation/widgets/log_in_widget.dart';
+export 'package:keno_plus/features/authentication/presentation/widgets/sign_up_widget.dart';
