+++ lib/features/authentication/presentation/bloc/authentication_event.dart
@@ -1,11 +1,31 @@
 part of 'authentication_bloc.dart';
 
-sealed class AuthenticationEvent {
-  const AuthenticationEvent();
+abstract class AuthenticationEvent {}
+
+class SwitchToLoginEvent extends AuthenticationEvent {}
+
+class SwitchToSignupEvent extends AuthenticationEvent {}
+
+class UpdateUserEmailEvent extends AuthenticationEvent {
+  final String userEmail;
+  UpdateUserEmailEvent(this.userEmail);
 }
 
-class OnCreatingUser extends AuthenticationEvent {
-  final User user;
-
-  OnCreatingUser({required this.user});
+class UpdatePasswordEvent extends AuthenticationEvent {
+  final String password;
+  UpdatePasswordEvent(this.password);
 }
+
+class UpdateUserEmailFocusEvent extends AuthenticationEvent {
+  final bool isFocused;
+  UpdateUserEmailFocusEvent(this.isFocused);
+}
+
+class UpdatePasswordFocusEvent extends AuthenticationEvent {
+  final bool isFocused;
+  UpdatePasswordFocusEvent(this.isFocused);
+}
+
+class TogglePasswordVisibilityEvent extends AuthenticationEvent {}
+
+class ResetPasswordVisibilityEvent extends AuthenticationEvent {}
